初回ログイン時の読み込み修正&CustomCell&Realm修正

+++ DayOfChoice/Model/Article.swift
@@ -53,7 +53,7 @@ public struct Friends: Codable {
 
 public struct ChartData {
     let title: String
-    let number: Int
+    let number: Double
     let color: Color
 }
